Move test_initial to tests/storage
